+++ crates/torii/typed-data/src/lib.rs
@@ -1,5 +1,8 @@
 #![warn(unused_crate_dependencies)]
 
+#[cfg(test)]
+mod test;
+
 pub mod error;
 pub mod typed_data;
